backend/api/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
faiss-cpu
langchain
openai
pymongo
langsmith
langchain-openai
numpy

